requirements.txt
fastapi
uvicorn
opencv-python
numpy
pytesseract
ultralytics
pillow
asyncio
requests
